adjustable smoothing, calculation caching
cbugg: closes bug-832

Should find a way to put this on the UI, but for now it is done with the
keyboard. 'p' increases smoothing, 'o' decreases.

As part of this, we no longer maintain a set of connected channels, but
just use the lamina permanent channel to do broadcasts.

A new session data object has been created that is also manipulated over
websocket. It is similar to the 'status' object, but is only changed by
the frontend, whereas 'status' is only changed by the backend.

This is where things that are adjustable during a session shall be kept,
like the smoothing window. In addition to the server using these
settings during calculation they're also broadcast to any connected
frontends to notify them to update their views.